Improve constants.h, adding support for larger data types
- Make the data properties automatically generated from the type

- Split DATA_MAX_SIZE into INPUT_DATA_MAX_WIDTH and INPUT_DATA_MAX_HEIGHT

- Add a height check into the InputParser

- Get rid of the shader compilation printout in the GPU generator

- Add support for different data types to the GPU generator

- Make the height overflow printout ellipsis scale with data size

- Use a root constant to define the data max value in the compute shaders instead of hardcoding

- Add a const value for the thread group size, and increase it to 64 for better performance on AMD hardware

+++ SummedAreaTableGeneratorCpuImpl.cpp
@@ -17,7 +17,7 @@ float SummedAreaTableGeneratorCpuImpl::generate(const DataContainer& data_in, Da
 	{
 		for (int x = 0; x < data_in.width; x++)
 		{
-			int output_value = data_in.data[y * data_in.width + x];
+			uint64_t output_value = data_in.data[y * data_in.width + x];
 
 			if (x > 0)
 			{
